+++ gw2buildutil/api/storage.py
@@ -37,10 +37,6 @@ class Storage (abc.ABC):
         pass
 
     @abc.abstractmethod
-    def exists (self, entity_type, api_id):
-        pass
-
-    @abc.abstractmethod
     def from_id (self, entity_type, id_):
         pass
 
@@ -55,6 +51,7 @@ class Storage (abc.ABC):
 
 class FileStorage (Storage):
     _SCHEMA_VERSION_KEY = 'meta:version'
+    _CONFLICT_VALUE = ''
 
     def __init__ (self, path=None):
         if path is None:
@@ -92,17 +89,17 @@ class FileStorage (Storage):
         else:
             return None
 
-    def _raw_key (self, path, api_id):
+    def _api_id_key (self, path, api_id):
         return f'entity:{"/".join(path)}:{api_id}'
 
     def store_raw (self, path, result):
-        self._raw_db[self._raw_key(path, result['id'])] = json.dumps(result)
+        self._raw_db[self._api_id_key(path, result['id'])] = json.dumps(result)
 
     def exists_raw (self, path, api_id):
-        return self._raw_key(path, api_id) in self._raw_db
+        return self._api_id_key(path, api_id) in self._raw_db
 
     def raw (self, path, api_id):
-        return json.loads(self._raw_db[self._raw_key(path, api_id)])
+        return json.loads(self._raw_db[self._api_id_key(path, api_id)])
 
     def clear_raw (self):
         for key in self._raw_db.keys():
@@ -112,15 +109,9 @@ class FileStorage (Storage):
         return (f'{entity_type.type_id()}:'
                 f'id:{util.Identified.normalise_id(id_)}')
 
-    def _api_id_key (self, entity_type, api_id):
-        return f'{entity_type.type_id()}:api_id:{api_id}'
-
     def store (self, entity):
         entity_type = type(entity)
-        data = json.dumps(entity.to_data())
-
-        # use latest if conflict
-        self._db[self._api_id_key(entity_type, entity.api_id)] = data
+        data = str(entity.api_id)
 
         for id_ in entity.ids:
             id_key = self._id_key(entity_type, id_)
@@ -131,25 +122,23 @@ class FileStorage (Storage):
             alias_key = self._id_key(entity_type, id_)
             # don't store if conflict
             if alias_key in self._db:
-                self._db[alias_key] = ''
+                if (self._db[alias_key].decode() not in
+                    (self._CONFLICT_VALUE, data)
+                ):
+                    self._db[alias_key] = self._CONFLICT_VALUE
             else:
                 self._db[alias_key] = data
 
-    def exists (self, entity_type, api_id):
-        return self._api_id_key(entity_type, api_id) in self._db
-
-    def _from_key (self, entity_type, key):
-        data = self._db[key]
-        if not data: # multiple entities used the same key
-            raise KeyError(key)
-        return entity_type.from_data(self, json.loads(data))
+    def from_api_id (self, entity_type, api_id):
+        return entity_type.from_api(self.raw(entity_type.path(), api_id), self)
 
     def from_id (self, entity_type, id_):
-        return self._from_key(entity_type, self._id_key(entity_type, id_))
-
-    def from_api_id (self, entity_type, api_id):
-        return self._from_key(entity_type,
-                              self._api_id_key(entity_type, api_id))
+        key = self._id_key(entity_type, id_)
+        api_id = self._db[key].decode()
+        if api_id == self._CONFLICT_VALUE:
+            # multiple entities used the same key
+            raise KeyError((entity_type, id_))
+        return self.from_api_id(entity_type, api_id)
 
     def clear (self):
         for key in self._db.keys():